# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TourWorld API
  description: |
    TourWorld 여행 관리 시스템 API

    ## 주요 기능
    - **일정표 관리 (FEAT-2)**: 여행 일정표 생성, 편집, 관리
    - **호텔 관리 (FEAT-1)**: 호텔 정보 등록 및 관리
    - **이미지 갤러리 (FEAT-3)**: 이미지 업로드 및 카테고리 관리

    ## 인증
    JWT 토큰 기반 인증을 사용합니다.
    - `Authorization: Bearer <token>` 헤더를 통해 인증

    ## 에러 응답
    모든 에러는 동일한 형식으로 반환됩니다:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "에러 메시지",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: TourWorld Team
servers:
  - url: http://localhost:3000/api
    description: 개발 서버
  - url: https://api.tourworld.com/api
    description: 프로덕션 서버
...
paths:
...
  /itineraries/{id}/items:
    get:
      tags:
        - ItineraryItems
      summary: 일정 항목 목록 조회
      operationId: listItineraryItems
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: dayNumber
          in: query
          schema:
            type: integer
          description: 특정 일차 필터
      responses:
        '200':
          description: 일정 항목 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListItineraryItemsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - ItineraryItems
      summary: 일정 항목 생성
      operationId: createItineraryItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItineraryItemRequest'
      responses:
        '201':
          description: 일정 항목 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateItineraryItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /itineraries/{id}/items/{itemId}:
    get:
      tags:
        - ItineraryItems
      summary: 일정 항목 상세 조회
      operationId: getItineraryItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/ItemIdParam'
      responses:
        '200':
          description: 일정 항목 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItineraryItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - ItineraryItems
      summary: 일정 항목 수정
      operationId: updateItineraryItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/ItemIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItineraryItemRequest'
      responses:
        '200':
          description: 일정 항목 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateItineraryItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - ItineraryItems
      summary: 일정 항목 삭제
      operationId: deleteItineraryItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/ItemIdParam'
      responses:
        '200':
          description: 일정 항목 삭제됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /itineraries/{id}/items/reorder:
    put:
      tags:
        - ItineraryItems
      summary: 일정 항목 순서 변경
      operationId: reorderItineraryItems
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderItineraryItemsRequest'
      responses:
        '200':
          description: 순서 변경됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReorderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # Hotels (FEAT-1)
  # ==========================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 인증 토큰

  # ==========================================
  # Parameters
  # ==========================================
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 리소스 ID
    ItemIdParam:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 하위 리소스 ID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 페이지 번호
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 페이지당 항목 수
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: 입력값 검증에 실패했습니다.
              details:
                errors:
                  - field: title
                    message: 필수 필드입니다.
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: 인증이 필요합니다.
...
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 요청한 리소스를 찾을 수 없습니다.
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        field:
          type: string
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalCount
        - totalPages
        - hasNextPage
        - hasPreviousPage
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
    Coordinates:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
    DeleteResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            deleted:
              type: boolean
              enum: [true]
    ReorderResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            reordered:
              type: boolean
              enum: [true]

    # Health Check
...
    ItineraryItem:
      type: object
      required:
        - id
        - itineraryId
        - dayNumber
        - title
        - sortOrder
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        itineraryId:
          type: string
          format: uuid
        hotelId:
          type: string
          format: uuid
          nullable: true
        dayNumber:
          type: integer
          minimum: 1
        startTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          nullable: true
        endTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateItineraryItemRequest:
      type: object
      required:
        - dayNumber
        - title
      properties:
        hotelId:
          type: string
          format: uuid
          nullable: true
        dayNumber:
          type: integer
          minimum: 1
        startTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          nullable: true
        endTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          nullable: true
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        sortOrder:
          type: integer
        imageIds:
          type: array
          items:
            type: string
            format: uuid
    UpdateItineraryItemRequest:
      type: object
      properties:
        hotelId:
          type: string
          format: uuid
          nullable: true
        dayNumber:
          type: integer
          minimum: 1
        startTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          nullable: true
        endTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          nullable: true
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        sortOrder:
          type: integer
        imageIds:
          type: array
          items:
            type: string
            format: uuid
    ReorderItineraryItemsRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - id
              - sortOrder
            properties:
              id:
                type: string
                format: uuid
              sortOrder:
                type: integer
              dayNumber:
                type: integer
                minimum: 1
    ListItineraryItemsResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required:
            - items
            - meta
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ItineraryItem'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
    GetItineraryItemResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/ItineraryItem'
    CreateItineraryItemResponse:
      $ref: '#/components/schemas/GetItineraryItemResponse'
    UpdateItineraryItemResponse:
      $ref: '#/components/schemas/GetItineraryItemResponse'

    # Hotel Schemas
...